refactor: FEAT-FEC-005 — unanimous_votes auf uint8_t einschränken (REQ-FEC-019)
Typ von uint32_t auf uint8_t geändert, da der Wertebereich 0..48 semantisch
korrekt auf uint8_t passt und im ALEWord-Struct 8 Byte Padding einspart.
Betroffen: ALEWord::unanimous_votes, decode_word_with_voting() Rückgabetyp
sowie lokale Variable und Test-Variable.

+++ include/FSK/symbol_decoder.h
@@ -67,8 +67,8 @@ static uint8_t bin_to_symbol(uint32_t bin_index);
      * \return             Number of unanimous votes among the 48 voted bit positions
      *                     (A.5.2.6.3); range 0..48.  Threshold: VOTE_THRESHOLD_BAD.
      */
-    static uint32_t decode_word_with_voting(const uint8_t symbols[],
-                                            uint64_t& output_word);
+    static uint8_t decode_word_with_voting(const uint8_t symbols[],
+                                           uint64_t& output_word);
     
 private:
     // Lookup table: FFT bin -> symbol value (not used anymore, kept for API compatibility)
